+++ resources/install.rb
@@ -20,16 +20,21 @@
 resource_name :hab_install
 
 property :install_url, String, default: 'https://raw.githubusercontent.com/habitat-sh/habitat/master/components/hab/install.sh'
-property :depot_url, String
+property :bldr_url, String
 property :version, String
 property :channel, String
 
 action :install do
+  if new_resource.version
+    Chef::Log.warn("Do not specify a version of Habitat with the 'hab_install' resource, it is ignored.")
+    Chef::Log.warn("The version property of the 'hab_install' resource will be removed in a future version.")
+    Chef::Log.warn("See https://github.com/chef-cookbooks/habitat/blob/master/README.md#habitat")
+  end
+
   if ::File.exist?(hab_path)
-    return unless new_resource.version
-    cmd = shell_out!([hab_path, '--version'])
+    cmd = shell_out!([hab_path, '--version', '0.33.2'])
     version = %r{hab (\d*\.\d*\.\d[^\/]*)}.match(cmd.stdout)[1]
-    return if new_resource.version == version
+    return if version == '0.33.2'
   end
 
   remote_file ::File.join(Chef::Config[:file_cache_path], 'hab-install.sh') do
@@ -38,7 +43,7 @@ action :install do
 
   execute 'installing with hab-install.sh' do
     command hab_command
-    environment 'HAB_DEPOT_URL' => new_resource.depot_url if new_resource.depot_url
+    environment 'HAB_BLDR_URL' => new_resource.bldr_url if new_resource.bldr_url
   end
 end
 
@@ -49,7 +54,7 @@ action :upgrade do
 
   execute 'installing with hab-install.sh' do
     command hab_command
-    environment 'HAB_DEPOT_URL' => new_resource.depot_url if new_resource.depot_url
+    environment 'HAB_BLDR_URL' => new_resource.bldr_url if new_resource.bldr_url
   end
 end
 
@@ -67,9 +72,7 @@ action_class do
   end
 
   def hab_command
-    cmd = ["bash #{Chef::Config[:file_cache_path]}/hab-install.sh"]
-    cmd.push("-v #{new_resource.version}") if new_resource.version
-    cmd.push("-c #{new_resource.channel}") if new_resource.channel
+    cmd = ["bash #{Chef::Config[:file_cache_path]}/hab-install.sh", '-v 0.33.2']
     cmd.join(' ')
   end
 end