+++ spec/unit/package_spec.rb
@@ -25,7 +25,7 @@ describe 'test::package' do
 
     it 'installs core/hab-sup with a specific depot url' do
       expect(chef_run).to install_hab_package('core/hab-sup')
-        .with(depot_url: 'http://app.acceptance.habitat.sh/v1/depot')
+        .with(bldr_url: 'http://app.acceptance.habitat.sh')
     end
   end
 end