+++ test/fixtures/cookbooks/test/recipes/install_for_chefspec.rb
@@ -2,10 +2,6 @@
 # not be used in the other recipes.
 hab_install 'install habitat'
 
-hab_install 'install habitat with version' do
-  version '0.12.0'
-end
-
 hab_install 'install habitat with depot url' do
-  depot_url 'https://localhost/v1/depot'
+  bldr_url 'https://localhost'
 end