+++ test/fixtures/cookbooks/test/recipes/service.rb
@@ -1,5 +1,4 @@
 hab_sup 'default' do
-  hab_version '0.26.1'
   hab_channel 'stable'
 end