+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file is used to list changes made in each version of the habitat cookbook.
 
+## 0.33.0 (2017-09-25)
+
+**Breaking change**
+
+This version is a compatibility change. All users of this cookbook must upgrade the cookbook to use current versions of Habitat.
+
+- Fix to account for [habitat-sh/habitat#3239](https://github.com/habitat-sh/habitat/pull/3239) - do not provide `/v1/depot` to Depot URL.
+- Pin the version of this cookbook along with the version of Habitat we install. This should match minor, but not necessarily patch versions.
+
 ## 0.28.0 (2017-09-19)
 
 - Add channel support to sup load/start